knot-mock - context mock
