+++ knotx-example/knotx-mocks/src/main/java/com/cognifide/knotx/mocks/knot/MockKnotHandler.java
@@ -25,22 +25,18 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import io.vertx.core.Handler;
-import io.vertx.core.file.OpenOptions;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.web.RoutingContext;
-import io.vertx.rxjava.core.buffer.Buffer;
 import io.vertx.rxjava.core.eventbus.Message;
-import io.vertx.rxjava.core.file.AsyncFile;
 import io.vertx.rxjava.core.file.FileSystem;
 import rx.Observable;
 
 public class MockKnotHandler implements Handler<Message<JsonObject>> {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(RoutingContext.class);
-  private static final OpenOptions OPEN_OPTIONS = new OpenOptions().setCreate(false).setWrite(false);
 
   private static final JsonObject ERROR_RESPONSE = new JsonObject().put(KnotContextKeys.RESPONSE.key(), new JsonObject().put("statusCode", 500));
   private static final JsonObject NOT_FOUND = new JsonObject().put(KnotContextKeys.RESPONSE.key(), new JsonObject().put("statusCode", 404));
@@ -89,68 +85,13 @@ public class MockKnotHandler implements Handler<Message<JsonObject>> {
     final JsonObject responseConfig = handledMocks.get(request.path());
 
     return Observable.from(KnotContextKeys.values())
-        .map(key -> mockData(key, responseConfig))
+        .map(key -> key.valueOrDefault(fileSystem, responseConfig))
+        .filter(value -> value.getRight() != null)
         .reduce(handlerResponse, this::mergeResponseValues);
   }
 
   private JsonObject mergeResponseValues(JsonObject result, Pair<String, Object> value) {
     return new JsonObject().put(value.getLeft(), value.getRight());
-  }
-
-  private Pair<String, Object> mockData(KnotContextKeys key, JsonObject responseConfig) {
-    final JsonObject value = new JsonObject();
-    if (responseConfig.containsKey(key.key())) {
-      value.put(key.key(), getMockFile(responseConfig.getString(key.key())));
-    } else {
-      value.put(key.key(), key.defaultValue());
-    }
-
-    return Pair.of(key.key(), value);
-  }
-
-  private Observable<JsonObject> getMockFile(String resourcePath) {
-    return fileSystem.openObservable(resourcePath, OPEN_OPTIONS)
-        .flatMap(this::processFile)
-        .map(Buffer::toJsonObject);
-  }
-
-  private Observable<Buffer> processFile(final AsyncFile asyncFile) {
-    return Observable.just(Buffer.buffer())
-        .mergeWith(asyncFile.toObservable())
-        .reduce(Buffer::appendBuffer);
-  }
-
-  private enum KnotContextKeys {
-    RESPONSE("clientResponse") {
-      @Override
-      Object defaultValue() {
-        return new JsonObject().put("statusCode", 200);
-      }
-    },
-    REQUEST("clientRequest") {
-      @Override
-      Object defaultValue() {
-        return new JsonObject();
-      }
-    },
-    FRAGMENTS("fragments") {
-      @Override
-      Object defaultValue() {
-        return new JsonArray();
-      }
-    };
-
-    private final String key;
-
-    KnotContextKeys(String key) {
-      this.key = key;
-    }
-
-    String key() {
-      return key;
-    }
-
-    abstract Object defaultValue();
   }
 
 }
