+++ knotx-example/knotx-mocks/src/main/resources/knotx-mocks.json
@@ -27,12 +27,14 @@
             "mocks": [
               {
                 "request.path": "/simple",
-                "fragments": "mock/knot/action/simple.fragments.json"
+                "fragments": "mock/knot/action/simple.fragments.json",
+                "transition": "knot.view"
               },
               {
                 "request.path": "/form-validation-error",
                 "fragments": "mock/knot/action/form-validation-error.fragments.json",
-                "clientResponse": "mock/knot/action/form-validation-error.response.json"
+                "clientResponse": "mock/knot/action/form-validation-error.response.json",
+                "transition": "knot.view"
               },
               {
                 "request.path": "/redirect",
